+++ lib/usb_cracker.rb
@@ -29,8 +29,10 @@ end # module UsbCracker
 
 require 'usb_cracker/candidates'
 require 'usb_cracker/cli'
+require 'usb_cracker/diskutil'
 require 'usb_cracker/prefix'
 require 'usb_cracker/secret_buffer'
+require 'usb_cracker/unlock'
 require 'usb_cracker/version'